Try to address the "Too many invisible frames" issues with containers.
By:
1. assembling encoded frames into super frame in SVT-VP9 plugin, to
   work around the super frame merging issue by VP9 super frame BSF;
2. splitting the merged super frames into individual ones by VP9
   super frame split BSF.

The SVT-VP9 encoder is designed to append 4 one-byte repeated frames
during Packetization, "periodically" determined by generate_rps_info()
in PictureDecision. But the 4 repeated frames are packetized with their
previous invisible frame as one (AVPacket) packet, which can't be parsed
as repeated frames when decoding them. So we have to separate the glued
frames and merge them as a super frame when receiving a packet in SVT-VP9
plugin, split the super frame (to resolve the issue that the repeated
frames point to the same invisible frame although their
index_of_frame_to_show are correct in the encoded bit stream), and
progressively increase the PTS/DTS of each frame.

Note:
1. the "-bsf:v vp9_superframe_split" option should be used into ffmpeg
   command line;
2. playback of the encoded bit stream has some jitter (repeated frame)
   issue, which may be due to the incompletely super frame splitting,
   and needs more investigation.

Signed-off-by: Austin Hu <austin.hu@intel.com>