Specific Resolution Fix
Fixed an issue that arose when frame sizes were non standard sized.
(e.g. non multiples of the largest coding unit (lcu) (64x64)).
For the example below the frame size used is 104x64. Encoded frames
showed a visible artifact.  In other cases, the frame could not be
successfully decoded by VQ Analyzer.

Root cause of the problem was in the logic that determined if the
current block to be coded contained a valid block (one which is fully
inside the frame).  If it did not contain a valid block, the current
block would be considered invalid and thus not coded.

The previous logic only checked the current block and blocks at
the next lowest depth for valid blocks.  However for certain sizes
(e.g. 104x64) some blocks require that a lower depth be searched
for valid blocks.

In this example, the 32x32 block at pixel location (96,0) was declared
invalid since itself and the 4 16x16 blocks it covers were all
invalid:
(96,0) 16x16 - invalid
(112,0) 16x16 - invalid
(96,16) 16x16- invalid
(112,16) 16x16 - invalid

However at the 8x8 depth there are indeed blocks that are valid:
(96,0) 8x8 - valid
(96,8) 8x8 - valid
(96,16) 8x8 - valid
(96,24) 8x8 - valid
etc...

One possible solution would be to expand the logic in the entropy
coding logic to recursively check blocks at lower levels, however
this might be too messy.  Also since this logic is executed
for every superblock on every frame, a better solution would be
to predetermine a valid block index for ever block in the superblock
once and store the results in an array.

Added a ep_scan_block_valid_block array similar to the
ep_scan_block_validity array in the SbParameters structure.

This array gets populated when the super block is initialized in
eb_vp9_sb_params_init.  The logic to initialize the array is to
iterate through the blocks (smallest to largest). The
valid block index for the current block is determined by examining only
that block and the 4 blocks it covers at a single lower depth.  If any
of those blocks are valid, then that block is recorded for the valid
block at the current block.

Predetermining the valid blocks should be a bit more efficient allowing
for a single array lookup, rather than a iterating through a loop
with 5 (ep_inter_depth_offset) comparison checks.

This PR fixes: https://github.com/OpenVisualCloud/SVT-VP9/issues/85
Incorrect encoded bitstream with some specific resolutions (such as
360x360)

Signed-off-by: Mark Feldman <mark.feldman@intel.com>